# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Address Normalizer API",
...
        "contact": {
            "name": "Address Normalizer Support",
            "url": "https://postalcodes.lumengroup.eu"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://postalcodes.lumengroup.eu",
            "description": "Production"
        },
        {
            "url": "http://localhost:8889",
            "description": "Local development"
        }
    ],
    "paths": {
...
        "/api/v1/validate": {
            "post": {
                "tags": [
                    "Address Validation"
                ],
                "summary": "Validate an address via Google Address Validation API",
                "description": "Validates an address directly through Google Address Validation API **without AI normalization**.\n\n**When to use:**\n- After a previous `/normalize` call (without Google validation) to verify the result\n- To get geographic coordinates for an already-clean address\n- To check address deliverability without running the full normalization pipeline\n\n**What it returns:** geographic coordinates (lat/lng), validation granularity, quality flags, property type (residential/business), and a list of issues found.\n\n**Note:** This endpoint requires Google Address Validation to be configured on the server (`GOOGLE_VALIDATION_API_KEY`). Returns `503` if not configured.",
                "operationId": "0b426d7229fa29f066ee34d0a22fc88b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "country",
                                    "city",
                                    "address"
                                ],
                                "properties": {
                                    "country": {
                                        "description": "ISO 3166-1 alpha-2 country code",
                                        "type": "string",
                                        "maxLength": 2,
                                        "minLength": 2,
                                        "example": "PL"
                                    },
                                    "postal_code": {
                                        "description": "Postal/ZIP code",
                                        "type": "string",
                                        "example": "00-001",
                                        "nullable": true
                                    },
                                    "city": {
                                        "description": "City name",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Warszawa"
                                    },
                                    "address": {
                                        "description": "Street address line",
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Marszałkowska 1/2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address validated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "ok"
                                            ],
                                            "example": "ok"
                                        },
                                        "validation": {
                                            "$ref": "#/components/schemas/GoogleValidationData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The country field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "$ref": "#/components/responses/RateLimitExceeded"
                    },
                    "503": {
                        "description": "Google Address Validation is not configured or unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Google Address Validation is not configured. Set GOOGLE_VALIDATION_API_KEY in server configuration."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/status": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "Health check and usage statistics",
                "description": "Returns the current API client status.\n\n**When to use:** To monitor your usage, check remaining quota before a large batch, or verify the API is operational.\n\n**Returns:** client name, monthly limit, current usage, remaining requests, cache hit rate (0.0–1.0), and which AI provider is active (openai or anthropic).",
                "operationId": "3f4b77f50fda21eafd0c314edb535673",
                "responses": {
                    "200": {
                        "description": "Client status and statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "client": {
                                            "description": "Client name",
                                            "type": "string",
                                            "example": "sportello"
                                        },
                                        "plan_limit": {
                                            "description": "Monthly request limit",
                                            "type": "integer",
                                            "example": 10000
                                        },
                                        "used_this_month": {
                                            "type": "integer",
                                            "example": 1234
                                        },
                                        "remaining": {
                                            "type": "integer",
                                            "example": 8766
                                        },
                                        "cache_hit_rate": {
                                            "description": "Ratio 0.0–1.0",
                                            "type": "number",
                                            "format": "float",
                                            "example": 0.72
                                        },
                                        "active_provider": {
                                            "type": "string",
                                            "enum": [
                                                "openai",
                                                "anthropic"
                                            ],
                                            "example": "openai"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "GoogleValidationData": {
                "title": "Google Validation Result",
                "description": "Address validation and geocoding data from Google Address Validation API. Only present when Google Address Validation is enabled on the server.",
                "properties": {
                    "latitude": {
                        "description": "Geographic latitude",
                        "type": "number",
                        "format": "double",
                        "example": 52.2297,
                        "nullable": true
                    },
                    "longitude": {
                        "description": "Geographic longitude",
                        "type": "number",
                        "format": "double",
                        "example": 21.0122,
                        "nullable": true
                    },
                    "place_id": {
                        "description": "Google Place ID — can be used with Google Maps/Places API",
                        "type": "string",
                        "example": "ChIJAZ-GmmbMHkcR_NPqiCq-8HI",
                        "nullable": true
                    },
                    "validation_granularity": {
                        "description": "How precisely Google matched the address. SUB_PREMISE (apartment level) is the most precise, OTHER means no match.",
                        "type": "string",
                        "enum": [
                            "OTHER",
                            "ROUTE",
                            "BLOCK",
                            "PREMISE_PROXIMITY",
                            "PREMISE",
                            "SUB_PREMISE"
                        ],
                        "example": "PREMISE"
                    },
                    "geocode_granularity": {
                        "description": "Precision level of the geographic coordinates",
                        "type": "string",
                        "enum": [
                            "OTHER",
                            "ROUTE",
                            "BLOCK",
                            "PREMISE_PROXIMITY",
                            "PREMISE",
                            "SUB_PREMISE"
                        ],
                        "example": "PREMISE"
                    },
                    "address_complete": {
                        "description": "All address components are present and confirmed by Google",
                        "type": "boolean",
                        "example": true
                    },
                    "has_unconfirmed_components": {
                        "description": "Some components could not be confirmed — may need manual review",
                        "type": "boolean",
                        "example": false
                    },
                    "has_inferred_components": {
                        "description": "Google filled in missing data (e.g. inferred postal code from city)",
                        "type": "boolean",
                        "example": false
                    },
                    "has_replaced_components": {
                        "description": "Google corrected/replaced some components (e.g. wrong postal code)",
                        "type": "boolean",
                        "example": false
                    },
                    "has_spell_corrected_components": {
                        "description": "Spelling corrections were applied to the address",
                        "type": "boolean",
                        "example": false
                    },
                    "is_residential": {
                        "description": "Address is classified as residential",
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "is_business": {
                        "description": "Address is classified as a business location",
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    },
                    "issues": {
                        "description": "Human-readable list of validation issues found (empty = no issues)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": []
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "API Bearer token authentication. Use: Bearer {your_token}",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}